[#15] fix: entity 필드 수정

+++ src/main/java/depromeet/lessonfour/server/activityrecords/domain/entities/Food.java
@@ -32,5 +32,5 @@ public class Food extends BaseTimeEntity {
   @NotNull private String name;
 
   @Column(nullable = false)
-  @NotNull private int foodScore;
+  @NotNull private double foodScore;
 }
